3-10  sanphams,sanpham-all,detail

+++ database/migrations/2025_09_18_005904_create_danh_gia_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('danh_gia', function (Blueprint $table) {
             $table->id();
-            $table->float('diem');
+            $table->tinyInteger('diem');
             $table->mediumText('noidung')->nullable();
             $table->mediumText('media')->nullable()->comment('phần nội dung đa phương tiện đi kèm với đánh giá (ảnh/video).');
             $table->datetime('ngaydang');
